refactor(core/acp): extract AcpStdioProcess base, port grok onto it
Pulls the shared ACP-over-stdio plumbing out of `core/providers/grok.ts`
into a new `core/providers/acp/base.ts`. No behavior change for the Grok
Build provider — the surface goes from 1116 to 435 lines because the
generic pieces (JSON-RPC pump, handshake, event translation, permission
flow, history injection) now live in the base class.

`AcpStdioProcess` is an abstract base implementing the full ACP wire
contract, parameterised by a small set of override hooks:

  • resolveSpawn(opts) — required, returns command + args + env
  • name — required, identifies the provider in logs
  • vendorNotificationPrefix — drops `_x.ai/*`, `cursor/*`, etc.
  • authenticate(opts, initResult) — optional second handshake step
  • preHandshake(opts) / onExit / onPreSpawnCleanup — bridge setup +
    teardown hooks
  • unwrapToolCall(update) — surfaces the canonical tool name behind
    vendor dispatch wrappers (Grok's `use_tool`, Cursor's `<kind>ToolCall`)
  • translateVendorUpdate(update) — short-circuit for vendor sub-types
  • pickPermissionOptionId, buildHistoryPromptBlock,
    initializeClientCapabilities, sessionNewParams — finer-grained
    knobs with sensible defaults

The shared logic that landed in the base includes everything we don't
want to re-implement per provider:

  • JSON-RPC pump (request/notification/server-request routing)
  • initialize + optional authenticate + session/new handshake
  • session/update → AgentEvent translation (5 standard sub-types)
  • session/request_permission bidirectional flow with bypassPermissions
    auto-approve
  • tool_call_update routing (input refresh vs result update)
  • Per-turn agent_message_chunk accumulation + synthetic `assistant`
    flush before `complete` (so persistence writes a row)
  • session/cancel interrupt
  • Pre-spawn race-safe kill()
  • History `resource` content block injection on the first turn
  • Backwards-compat `serializeHistoryForAcp` (re-exported from grok.ts
    as `serializeHistoryForGrok` for any external callers that used it)

GrokProcess now extends AcpStdioProcess and keeps only what's truly
Grok-specific: `resolveGrokPath`, the top-level flag ordering for
`grok agent stdio`, `_x.ai/*` prefix drop, the `use_tool` dispatch
unwrap (xAI's MCP wrapper exposes `rawInput.tool_name` / `tool_input`),
and the `<cwd>/.grok/config.toml` + stdio→HTTP MCP bridge setup.
GrokProvider's `complete()` (`grok -p` headless path) and `listModels`
stay in this file too — they're stateless and don't share code with
the ACP stdio path.

Regression-tested against grok 0.1.212: spawn → init → send →
agent_message_chunk stream → complete → kill round-trip remains
identical (live smoke run returned "Hi there friend" in ~5s, same as
before the refactor).
